[EXPERIMENTAL/BROKEN] try to implement partial transparency
As implemented in Notepad++, a partially transparent search/replace window could make it easier to interact with the editor (particularly in fullscreen contexts, when the "always on top" window could hide the search results).
Unfortunately, Java can only apply the transparency to undecorated window, so it's a no-go.
Java gurus, please come to the rescue :D